meson: simplify use of the common-main library
The "common-main.c" file is used by multiple executables. In order to
make it easy to set it up we have created a separate library that these
executables can link against. All of these executables also want to link
against `libgit.a` though, which makes it necessary to specify both of
these as dependencies for every executable.

Simplify this a bit by declaring the library as a source dependency:
instead of creating a static library, we now instead compile the common
set of files into each executable separately.

This change surfaces an issue when linking aliases for git-remote-http:
we extract all objects from `git-remote-http` et al and then link them
into the new executable. As such, these objects would already contain
a `main()` function. But now that we also compile "common-main.c" into
these aliased executables we see a linker error due to `main()` being
defined twice. We fix this by only linking against `libgit.a`.

Signed-off-by: Patrick Steinhardt <ps@pks.im>
Signed-off-by: Junio C Hamano <gitster@pobox.com>